perf(cli): extract cleanup command contract
