+++ crates/contracts/homeboy-command-contract/Cargo.toml
@@ -13,5 +13,6 @@ name = "homeboy_command_contract"
 path = "src/lib.rs"
 
 [dependencies]
+clap = { version = "4.5", features = ["derive", "string"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
